stat service is now working

+++ src/main/java/com/finalfantasy/football/api/StatRequest.java
@@ -1,6 +1,6 @@
 package com.finalfantasy.football.api;
 
 public class StatRequest {
-  public int week;
-  public int season;
+  public short week;
+  public short season;
 }
